Fix integers in collection assert
When validating a collection where one of the values is an
integer, the `validate()` method would throw because it only
handles strings, arrays or binded objects.

The error thrown was:

```
Error: You must validate an object or a string
```

In the case of an integer, we need it to fallback to validating
an "object", which is the default behavior. Also, removing this
does not cause any of our errors to fail.

At the moment I don't see any downsides to this. Thoughts?